Update people view to get email from emails table
The column is there but empty (no user has it set in the production DB).
It turns out it is stored in the `emails` table. This supports users
having multiple emails, I suspect to enable a single user having access
to multiple communities.

We don't care about this since there's only a community for Donalo so
we can safely assume a one to one relationship between person and email
records.