Fix listing show due to not informed custom fields
That's because as the custom fields "Cantidad mínima a solicitar" and
"Cantidad disponible" didn't have an associated category other than the
default one (the first one), the listing/new view didn't show the fields
unless selecting the default category. As a result, the listing had no
custom field values associated and failed to load them on :show.